Module aws_sdk_kinesisvideoarchivedmedia::error

Modules

client_limit_exceeded_exception

See ClientLimitExceededException

invalid_argument_exception

See InvalidArgumentException

invalid_codec_private_data_exception

See InvalidCodecPrivateDataException

invalid_media_frame_exception

See InvalidMediaFrameException

missing_codec_private_data_exception

See MissingCodecPrivateDataException

no_data_retention_exception

See NoDataRetentionException

not_authorized_exception

See NotAuthorizedException

resource_not_found_exception

See ResourceNotFoundException

unsupported_stream_media_type_exception

See UnsupportedStreamMediaTypeException

Structs

ClientLimitExceededException

Kinesis Video Streams has throttled the request because you have exceeded a limit. Try making the call later. For information about limits, see Kinesis Video Streams Limits.

GetClipError

GetDASHStreamingSessionURLError

GetHLSStreamingSessionURLError

GetMediaForFragmentListError

InvalidArgumentException

A specified parameter exceeds its restrictions, is not supported, or can't be

InvalidCodecPrivateDataException

The codec private data in at least one of the tracks of the video stream is not valid

InvalidMediaFrameException

One or more frames in the requested clip could not be parsed based on the specified

ListFragmentsError

MissingCodecPrivateDataException

No codec private data was found in at least one of tracks of the video stream.

NoDataRetentionException

A streaming session was requested for a stream that does not retain data (that is, has

NotAuthorizedException

Status Code: 403, The caller is not authorized to perform an operation on the given

ResourceNotFoundException

UnsupportedStreamMediaTypeException

The type of the media (for example, h.264 or h.265 video or ACC or G.711 audio) could

Enums

GetClipErrorKind

GetDASHStreamingSessionURLErrorKind

GetHLSStreamingSessionURLErrorKind

GetMediaForFragmentListErrorKind

ListFragmentsErrorKind