Module aws_sdk_kinesisvideoarchivedmedia::model

Modules

clip_fragment_selector

See ClipFragmentSelector

clip_timestamp_range

See ClipTimestampRange

dash_fragment_selector

See DashFragmentSelector

dash_timestamp_range

See DashTimestampRange

fragment

See Fragment

fragment_selector

See FragmentSelector

hls_fragment_selector

See HlsFragmentSelector

hls_timestamp_range

See HlsTimestampRange

timestamp_range

See TimestampRange

Structs

ClipFragmentSelector

Describes the timestamp range and timestamp origin of a range of fragments.

ClipTimestampRange

The range of timestamps for which to return fragments.

DashFragmentSelector

Contains the range of timestamps for the requested media, and the source of the

DashTimestampRange

The start and end of the timestamp range for the requested media.

Fragment

Represents a segment of video or other time-delimited data.

FragmentSelector

Describes the timestamp range and timestamp origin of a range of fragments.

HlsFragmentSelector

Contains the range of timestamps for the requested media, and the source of the

HlsTimestampRange

The start and end of the timestamp range for the requested media.

TimestampRange

The range of timestamps for which to return fragments.

Enums

ClipFragmentSelectorType

ContainerFormat

DashDisplayFragmentNumber

DashDisplayFragmentTimestamp

DashFragmentSelectorType

DashPlaybackMode

FragmentSelectorType

HlsDiscontinuityMode

HlsDisplayFragmentTimestamp

HlsFragmentSelectorType

HlsPlaybackMode